Исправлено восстановление и замена пароля для админа и юзера #308
